Cope with function deletion and re-creation in close succession
`riff update` current provokes the behaviour, but it could happen if a function
was explicitly deleted and then immediated re-created.

The approach is to retry on function-controller's main goroutine. This will
make the main goroutine unresponsive.

An alternative would be to spawn a goroutine to retry and complete the
deployment, but this state would need to be tracked to handle related activity
correctly. This seemed overly complex given function-controller's current
design of taking action on the main goroutine.

Fixes https://github.com/projectriff/function-controller/issues/5